backend/requirements.txt
pymongo
dnspython
python-dotenv